% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longley.R
\docType{data}
\name{longley}
\alias{longley}
\title{Longley}
\format{A data frame with 16 rows and 7 colums
\itemize{
\item GNP.deflator range 80 - 120
\item GNP Gross National Product
\item Unemployed Unemployment rate
\item Armed.Forces
\item Population
\item Year
\item Employed
}}
\usage{
longley
}
\description{
Longley's Economic Regression Data
}
\author{
Danny Ettelson
}
\keyword{datasets}
